core domains modules merge (cargo + handling)
