+++ cargo/src/main/java/br/ufes/dwws/cargo/CargoRepository.java
@@ -5,4 +5,6 @@ import br.ufes.dwws.utils.ddd.Repository;
 public interface CargoRepository extends Repository<Cargo, TrackingId> {
 
     TrackingId nextId();
+
+    void save(Cargo cargo);
 }
