--- cargo/src/main/java/br/ufes/dwws/cargo/RoutingStatus.java
@@ -1,6 +1,0 @@
-package br.ufes.dwws.cargo;
-
-public enum RoutingStatus {
-
-    NOT_ROUTED, ROUTED, MISROUTED;
-}
