+++ cargo/src/main/java/br/ufes/dwws/handling/HandlingEventType.java
@@ -1,4 +1,4 @@
-package br.ufes.dwws.cargo;
+package br.ufes.dwws.handling;
 
 public enum HandlingEventType {
