+++ location/src/main/resources/META-INF/orm.xml
@@ -2,7 +2,7 @@
 <entity-mappings xmlns="https://jakarta.ee/xml/ns/persistence/orm" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://jakarta.ee/xml/ns/persistence/orm https://jakarta.ee/xml/ns/persistence/orm/orm_3_0.xsd" version="3.0">
 
     <package>br.ufes.dwws.location</package>
-    <schema>location</schema>
+    <!--<schema>location</schema>-->
     <access>FIELD</access>
 
     <!--
