+++ voyage/src/main/resources/META-INF/orm.xml
@@ -2,7 +2,7 @@
 <entity-mappings xmlns="https://jakarta.ee/xml/ns/persistence/orm" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://jakarta.ee/xml/ns/persistence/orm https://jakarta.ee/xml/ns/persistence/orm/orm_3_0.xsd" version="3.0">
 
     <package>br.ufes.dwws.voyage</package>
-    <schema>voyage</schema>
+    <!--<schema>voyage</schema>-->
     <access>FIELD</access>
 
     <!--
@@ -27,11 +27,7 @@
     </entity>
     -->
 
-    <embeddable class="VoyageNumber" metadata-complete="true">
-        <attributes>
-            <basic name="number" optional="false"/>
-        </attributes>
-    </embeddable>
+    <embeddable class="VoyageNumber" metadata-complete="true"/>
 
     <embeddable class="CarrierMovement" metadata-complete="true">
         <attributes>
